requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.2
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openpyxl==3.1.2
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
yarl==1.9.2
